feat: move portfolio data fetch to content
